LU-9078 lnet: Fix route hops print
The default number of hops for  a route is -1. This is
currently being printed as %u. Change that to %d to
make it print out properly.

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/25250
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lnet/utils/lnetconfig/liblnetconfig.c
@@ -807,8 +807,8 @@ int lustre_lnet_show_route(char *nw, char *gw, int hops, int prio, int detail,
 
 		if (detail) {
 			if (cYAML_create_number(item, "hop",
-						data.cfg_config_u.cfg_route.
-							rtr_hop) ==
+						(int) data.cfg_config_u.
+						cfg_route.rtr_hop) ==
 			    NULL)
 				goto out;
